Split coffee management into green and roasted modules
