+++ src/models/greenCoffeeData.js
@@ -1,0 +1,19 @@
+import Coffee from './coffee';
+
+const greenCoffeeData = [
+  new Coffee({
+    name: 'Guatemala Huehuetenango',
+    origins: ['Guatemala'],
+    providerId: 'provider1',
+    producer: 'Juan Perez',
+    farm: 'Finca La Esperanza',
+    region: 'Huehuetenango',
+    process: 'Washed',
+    varietal: 'Bourbon',
+    altitude: '1800m',
+    tastingNotes: 'Chocolate, Citrus',
+    purchasePrice: '5.00',
+  }),
+];
+
+export default greenCoffeeData;
