+++ src/models/modules.js
@@ -1,9 +1,10 @@
-import { FaBox, FaChartBar, FaChartLine, FaCoffee, FaCog, FaFileInvoiceDollar, FaTools } from 'react-icons/fa';
+import { FaBox, FaChartBar, FaChartLine, FaCoffee, FaCog, FaFileInvoiceDollar, FaTools, FaLeaf } from 'react-icons/fa';
 
 const modules = [
-  { key: 'coffee', name: 'Coffees', icon: FaCoffee, enabled: true },
-  { key: 'inventory', name: 'Inventory', icon: FaBox, enabled: true },
   { key: 'dashboard', name: 'Dashboard', icon: FaChartLine, enabled: true },
+  { key: 'greenCoffee', name: 'Green Coffee', icon: FaLeaf, enabled: true },
+  { key: 'roastedCoffee', name: 'Roasted Coffee', icon: FaCoffee, enabled: true },
+  { key: 'inventory', name: 'Inventory', icon: FaBox, enabled: true },
   { key: 'projections', name: 'Projections', icon: FaChartBar, enabled: true },
   { key: 'settings', name: 'Settings', icon: FaCog, enabled: true },
   { key: 'billing', name: 'Billing', icon: FaFileInvoiceDollar, enabled: true },
